move sphere tree handeling to multi sphere
